+++ src/main/java/roomescape/auth/sign/application/usecase/SignUpUseCase.java
@@ -1,9 +1,26 @@
 package roomescape.auth.sign.application.usecase;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.auth.sign.application.dto.CreateUserRequest;
+import roomescape.auth.sign.password.Password;
+import roomescape.auth.sign.password.PasswordEncoder;
 import roomescape.user.application.dto.SignUpRequest;
+import roomescape.user.application.service.UserCommandService;
 import roomescape.user.domain.User;
 
-public interface SignUpUseCase {
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class SignUpUseCase {
 
-    User execute(SignUpRequest request);
+    private final PasswordEncoder passwordEncoder;
+    private final UserCommandService userCommandService;
+
+    public User execute(final SignUpRequest request) {
+        final Password encodedPassword = Password.fromRaw(request.rawPassword(), passwordEncoder);
+
+        return userCommandService.create(CreateUserRequest.from(request, encodedPassword));
+    }
 }
