+++ src/main/java/roomescape/common/cookie/setter/CookieSetter.java
@@ -2,8 +2,19 @@ package roomescape.common.cookie.setter;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.stereotype.Component;
 
-public interface CookieSetter {
+@Component
+public class CookieSetter {
 
-    void execute(final HttpServletResponse response, Cookie cookie);
+    public void execute(final HttpServletResponse response,
+                        final Cookie cookie) {
+        if (response == null) {
+            throw new IllegalArgumentException("HTTP response cannot be null");
+        }
+        if (cookie == null) {
+            throw new IllegalArgumentException("cookie cannot be null");
+        }
+        response.addCookie(cookie);
+    }
 }
