+++ src/main/java/roomescape/time/application/service/ReservationTimeQueryService.java
@@ -1,18 +1,37 @@
 package roomescape.time.application.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.domain.ReservationTimeId;
+import roomescape.time.domain.ReservationTimeRepository;
 
 import java.time.LocalTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 
-public interface ReservationTimeQueryService {
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ReservationTimeQueryService {
 
-    ReservationTime get(ReservationTimeId id);
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    List<ReservationTime> getAll();
+    public ReservationTime get(final ReservationTimeId id) {
+        return reservationTimeRepository.findById(id)
+                .orElseThrow(NoSuchElementException::new);
+    }
 
-    boolean existsByStartAt(LocalTime time);
+    public List<ReservationTime> getAll() {
+        return reservationTimeRepository.findAll();
+    }
 
-    boolean existById(ReservationTimeId id);
+    public boolean existsByStartAt(final LocalTime startAt) {
+        return reservationTimeRepository.existsByStartAt(startAt);
+    }
+
+    public boolean existById(final ReservationTimeId id) {
+        return reservationTimeRepository.existsById(id);
+    }
 }
