+++ src/test/java/roomescape/auth/cookie/extractor/CookieExtractorTest.java
@@ -14,7 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest
-class CookieExtractorImplTest {
+class CookieExtractorTest {
 
     @Autowired
     private CookieExtractor cookieExtractor;
