+++ src/test/java/roomescape/auth/jwt/generator/JwtGeneratorTest.java
@@ -21,7 +21,7 @@ import java.util.Date;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
-class HmacJwtGeneratorTest {
+class JwtGeneratorTest {
 
     private static final SecretKey secretKey =
             Keys.hmacShaKeyFor("ZMQ5kFGf1MTmnkhik/7rOcmT6OPvlz5Z+4aP4pAfNtk=".getBytes(StandardCharsets.UTF_8));
