+++ src/test/java/roomescape/reservation/application/service/ReservationQueryServiceTest.java
@@ -33,10 +33,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 @SpringBootTest
 @Transactional
-class ReservationQueryServiceImplTest {
+class ReservationQueryServiceTest {
 
     @Autowired
-    private ReservationQueryServiceImpl reservationQueryService;
+    private ReservationQueryService reservationQueryService;
 
     @Autowired
     private ReservationRepository reservationRepository;
