+++ src/test/java/roomescape/time/application/service/ReservationTimeCommandServiceTest.java
@@ -35,10 +35,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest
 @Transactional
-class ReservationTimeCommandServiceImplTest {
+class ReservationTimeCommandServiceTest {
 
     @Autowired
-    private ReservationTimeCommandServiceImpl reservationTimeCommandService;
+    private ReservationTimeCommandService reservationTimeCommandService;
 
     @Autowired
     private ReservationTimeRepository reservationTimeRepository;
